DriveStraight and Spin to Both work ... Thanks Adrian
Added toggle method for the controller. adrian brought up a good point that if we can get drivestraight with gyro to work just fine, we should enable that drive assistmode all the time and allow it to be manually disabled. Instead of just having the PID output be the turn mod, ad the joysticks' output to the turn